Fix add-on absorbing input it doesn't use
For example, 2048 doesn't use most buttons, but
CControllerInput::GetInputType() doesn't know this, so it optimistically
returns the type defined by the controller's layout.xml instead of
INPUT_TYPE_UNKNOWN.

The correect fix for this would be to add a HasFeature() call to the Game
API, that returns true when the feature is handled by the add-on and false
otherwise.

NOTE: This only fixes digital buttons mapped to digital features. If a
digital button is mapped to an analog feature (such as left bumper being
mapped to left trigger), the feature will still be incorrectly absorbed.