Fix long freezes when calling CAddon::LibPath()
The freezes occur when retrieving the LibPath of uninstalled add-ons. In
this case, LibPath() is broken because it returns the path to the repo
archive instead of the add-on's install location. E.g. LibPath() might
look like:

https://googlecode.com/.../service.rom.collection.browser-1.0.0.zip/default.py

In this case, CFile::Exists() will freeze while checking the https URL.
The obvious fix is to only call Exists() if the add-on is installed to
the share path. However, there may be a better approach all together.